camera_calibration_ws/monodepth-FPN/MonoDepth-FPN-PyTorch/requirements.txt
absl-py==0.13.0
actionlib==1.12.1
angles==1.9.12
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base_local_planner==1.16.7
bleach==3.2.2
cachetools==4.2.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
certifi==2021.5.30
cffi==1.14.4
charset-normalizer==2.0.3
cv_bridge==1.13.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
diagnostic_updater==1.9.7
distro==1.5.0
docutils==0.16
dynamic_reconfigure==1.6.3
entrypoints==0.3
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
google-auth==1.34.0
google-auth-oauthlib==0.4.4
grpcio==1.39.0
idna==3.2
image_geometry==1.13.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.4.0
interactive-markers==1.11.5
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.2
jsk-recognition-utils==1.2.15
jsk_rviz_plugins==2.1.7
jsk_topic_tools==2.2.11
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
laser_geometry==1.6.7
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
message_filters==1.14.12
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
net-utils==1.3
networkx==2.5
notebook==6.2.0
numpy==1.19.5
nvidia-pyindex==1.0.9
oauthlib==3.1.1
open3d-python==0.7.0.0
opencv-python==4.5.1.48
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
path==15.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.17.3
python-dateutil==2.8.1
python_qt_binding==0.4.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==21.0.2
qt-gui==0.4.2
requests==2.26.0
requests-oauthlib==1.3.0
resource_retriever==1.12.6
ros_numpy==0.0.3
rosbag==1.14.12
rosclean==1.14.9
rosgraph==1.14.12
roslaunch==1.14.12
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.12
rosmaster==1.14.12
rosmsg==1.14.12
rosnode==1.14.12
rosparam==1.14.12
rospkg==1.2.10
rospy==1.14.12
rosservice==1.14.12
rostest==1.14.12
rostopic==1.14.12
rosunit==1.14.9
roswtf==1.14.12
rqt-reconfigure==0.5.4
rqt_bag==0.5.1
rqt_console==0.4.9
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_logger_level==0.4.8
rqt_py_common==0.5.2
rsa==4.7.2
rviz==1.13.21
scikit-image==0.17.2
scipy==1.2.0
Send2Trash==1.5.0
sensor-msgs==1.12.8
Shapely==1.7.1
six==1.15.0
smclib==1.8.5
sound-play==0.3.12
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.2
testpath==0.4.4
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
tifffile==2020.9.3
toml==0.10.2
topic_tools==1.14.12
torch==1.7.1+cu110
torch-tb-profiler==0.2.1
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urdfdom-py==0.4.6
urllib3==1.26.6
utils==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
